Calculation needs dbl variable [Findbugs]
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jmeter/branches/rel-2-1@408511 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 1f42b48f65fa151d54a359dcf952c7f51c1d9ebd

+++ src/core/org/apache/jmeter/util/Calculator.java
@@ -232,7 +232,7 @@ public class Calculator {
      */
     public double getPageSize() {
         if (count > 0 && bytes > 0) {
-            return bytes / count;
+            return (double) bytes / count;
         }
         return 0.0;
     }
